# src/main/resources/swagger/ms-creditos_swagger.yaml
openapi: 3.0.1
info:
  title: Microservicio de Créditos
  version: 1.0.0
  description: |-
    El Microservicio de Créditos es una API diseñada para gestionar y administrar créditos y movimientos relacionados con créditos. Proporciona endpoints para realizar operaciones como obtener información de créditos, insertar nuevos créditos, actualizar créditos existentes, eliminar créditos, obtener movimientos de créditos, crear nuevos movimientos, pagar cuotas de movimientos y más. Este microservicio es una herramienta poderosa para el seguimiento y control de préstamos y movimientos financieros asociados.
  termsOfService: https://www.example.com/terms
  contact:
    email: vertkle@gmail.com
  license:
    name: Licencia MIT
    url: https://www.example.com/license
servers:
  - url: http://localhost:8080/api/v1
paths:
  /credits:
    get:
      operationId: getAllCredits
      summary: Obtener todos los créditos
      tags:
        - Créditos
      responses:
        '200':
          description: Lista de créditos obtenida exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Credit'
    post:
      operationId: createCredit
      summary: Insertar un nuevo crédito
      tags:
        - Créditos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Credit'
      responses:
        '201':
          description: Crédito insertado exitosamente
        '400':
          description: Error en request
    put:
      operationId: updateCredit
      summary: Actualizar un crédito existente
      tags:
        - Créditos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Credit'
      responses:
        '200':
          description: Actualizado correctamente
        '400':
          description: Error en el request
        '404':
          description: Recurso no encontrado
  /credits/{credit_id}:
    get:
      operationId: getCreditById
      summary: Trae información de un crédito según su credit_id
      tags:
        - Créditos
      parameters:
        - in: path
          name: credit_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Crédito obtenido correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Credit'
        '400':
          description: Error en Request
        '404':
          description: Recurso no encontrado
    delete:
      operationId: deleteCreditById
      summary: Eliminar un crédito por ID
      tags:
        - Créditos
      parameters:
        - in: path
          name: credit_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Crédito eliminado exitosamente
        '400':
          description: Error en Request
        '404':
          description: Recurso no encontrado
  /credits/customer/{customer_id}:
    get:
      operationId: getCreditsByCustomerId
      summary: Obtener los créditos por ID de cliente
      tags:
        - Créditos
      parameters:
        - in: path
          name: customer_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Lista de créditos obtenida exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Credit'
...
components:
  schemas:
    Credit:
      type: object
      properties:
        _id:
          type: string
        customer_id:
          type: string
        type_credit:
          type: string
        mount_limit:
          type: number
        date_open:
          type: string
        interest_rate:
          type: number
...